bpmodels: Simulating and Analysing Transmission Chain Statistics using Branching Process Models

Provides methods to simulate and analyse the size and length of branching processes with an arbitrary offspring distribution. These can be used, for example, to analyse the distribution of chain sizes or length of infectious disease outbreaks, as discussed in Farrington et al. (2003) doi:10.1093/biostatistics/4.2.279 .

See also

Author

Maintainer: James M. Azam james.azam@lshtm.ac.uk (ORCID)

Authors:

Flavio Finger flavio.finger@epicentre.msf.org (ORCID)

Sebastian Funk sebastian.funk@lshtm.ac.uk (ORCID) [copyright holder]

Other contributors:

Zhian N. Kamvar zkamvar@gmail.com (ORCID) [contributor]