Simulate chains using a branching process

Simulate chains using a branching process

chain_sim(n, offspring, stat = c("size", "length"), infinite = Inf,
  tree = FALSE, serial, t0 = 0, tf = Inf, ...)

Arguments

n

number of simulations to run.

offspring

offspring distribution: a character string corresponding to the R distribution function (e.g., "pois" for Poisson, where rpois is the R function to generate Poisson random numbers)

stat

statistic to calculate ("size" or "length" of chains)

infinite

a size or length from which the size/length is to be considered infinite

tree

return the tree of infectors

serial

the serial interval; a function that takes one parameter (`n`), the number of serial intervals to randomly sample; if this parameter is set, `chain_sim` returns times of infection, too; implies (`tree`=TRUE)

t0

start time (if serial interval is given); either a single value (0 by default for all simulations, or a vector of length `n` with initial times)

tf

end time (if serial interval is given)

...

parameters of the offspring distribution

Value

a vector of sizes/lengths (if tree==FALSE and no serial interval given), or a data frame with columns `n` (simulation ID), `time` (if the serial interval is given) and (if tree==TRUE) `id` (a unique ID within each simulation for each individual element of the chain), `ancestor` (the ID of the ancestor of each element) and `generation`.

Author

Sebastian Funk

Examples

chain_sim(n=5, "pois", "size", lambda=0.5)
#> [1] 1 1 2 1 2